Add `create_bss_text_factory` helper function
