+++ gui/src/widgets/bss_table/columns/bssid.rs
@@ -1,35 +1,11 @@
 use gtk::SignalListItemFactory;
-use gtk::prelude::*;
 
-use super::set_bss_label;
-use crate::objects::BssObject;
+use super::{create_bss_sorter_by, create_bss_text_factory};
 
 pub fn create_bssid_factory() -> SignalListItemFactory {
-    let factory = SignalListItemFactory::new();
-
-    factory.connect_setup(move |_, list_item| {
-        let label = gtk::Label::new(None);
-        label.set_halign(gtk::Align::Start);
-        list_item
-            .downcast_ref::<gtk::ListItem>()
-            .unwrap()
-            .set_child(Some(&label));
-    });
-
-    factory.connect_bind(move |_, list_item| {
-        let list_item = list_item.downcast_ref::<gtk::ListItem>().unwrap();
-        let bss = list_item.item().and_downcast::<BssObject>().unwrap();
-        let label = list_item.child().and_downcast::<gtk::Label>().unwrap();
-        set_bss_label(&label, bss.bssid(), bss.is_associated());
-    });
-
-    factory
+    create_bss_text_factory(gtk::Align::Start, |bss| Some(bss.bssid()))
 }
 
 pub fn create_bssid_sorter() -> gtk::CustomSorter {
-    gtk::CustomSorter::new(|obj1, obj2| {
-        let bss1 = obj1.downcast_ref::<BssObject>().unwrap();
-        let bss2 = obj2.downcast_ref::<BssObject>().unwrap();
-        bss1.bssid().cmp(&bss2.bssid()).into()
-    })
+    create_bss_sorter_by(|bss| bss.bssid())
 }
