+++ gui/src/widgets/bss_table/columns/channel_width.rs
@@ -1,26 +1,10 @@
 use gtk::SignalListItemFactory;
-use gtk::prelude::*;
 use kawaiifi::ChannelWidth;
 
-use super::set_bss_label;
-use crate::objects::BssObject;
+use super::{create_bss_sorter_by, create_bss_text_factory};
 
 pub fn create_channel_width_factory() -> SignalListItemFactory {
-    let factory = SignalListItemFactory::new();
-
-    factory.connect_setup(move |_, list_item| {
-        let label = gtk::Label::new(None);
-        label.set_halign(gtk::Align::End);
-        list_item
-            .downcast_ref::<gtk::ListItem>()
-            .unwrap()
-            .set_child(Some(&label));
-    });
-
-    factory.connect_bind(move |_, list_item| {
-        let list_item = list_item.downcast_ref::<gtk::ListItem>().unwrap();
-        let bss = list_item.item().and_downcast::<BssObject>().unwrap();
-        let label = list_item.child().and_downcast::<gtk::Label>().unwrap();
+    create_bss_text_factory(gtk::Align::End, |bss| {
         let width = match bss.channel_width() {
             ChannelWidth::TwentyMhz => "20",
             ChannelWidth::FortyMhz => "40",
@@ -29,16 +13,10 @@ pub fn create_channel_width_factory() -> SignalListItemFactory {
             ChannelWidth::OneSixtyMhz => "160",
             ChannelWidth::ThreeHundredTwentyMhz => "320",
         };
-        set_bss_label(&label, format!("{width} MHz"), bss.is_associated());
-    });
-
-    factory
+        Some(format!("{width} MHz"))
+    })
 }
 
 pub fn create_channel_width_sorter() -> gtk::CustomSorter {
-    gtk::CustomSorter::new(|obj1, obj2| {
-        let bss1 = obj1.downcast_ref::<BssObject>().unwrap();
-        let bss2 = obj2.downcast_ref::<BssObject>().unwrap();
-        bss1.channel_width().cmp(&bss2.channel_width()).into()
-    })
+    create_bss_sorter_by(|bss| bss.channel_width())
 }
