+++ gui/src/widgets/bss_table/columns/security.rs
@@ -1,29 +1,11 @@
 use gtk::SignalListItemFactory;
 use gtk::prelude::*;
 
-use super::set_bss_label;
+use super::create_bss_text_factory;
 use crate::objects::BssObject;
 
 pub fn create_security_factory() -> SignalListItemFactory {
-    let factory = SignalListItemFactory::new();
-
-    factory.connect_setup(move |_, list_item| {
-        let label = gtk::Label::new(None);
-        label.set_halign(gtk::Align::Start);
-        list_item
-            .downcast_ref::<gtk::ListItem>()
-            .unwrap()
-            .set_child(Some(&label));
-    });
-
-    factory.connect_bind(move |_, list_item| {
-        let list_item = list_item.downcast_ref::<gtk::ListItem>().unwrap();
-        let bss = list_item.item().and_downcast::<BssObject>().unwrap();
-        let label = list_item.child().and_downcast::<gtk::Label>().unwrap();
-        set_bss_label(&label, bss.security().to_string(), bss.is_associated());
-    });
-
-    factory
+    create_bss_text_factory(gtk::Align::Start, |bss| Some(bss.security().to_string()))
 }
 
 pub fn create_security_sorter() -> gtk::CustomSorter {
