+++ gui/src/widgets/bss_table/mod.rs
@@ -1,19 +1,19 @@
 mod columns;
+mod setup;
 
 use std::cell::RefCell;
 use std::rc::Rc;
 use std::time::Duration;
 
-use gtk::gio::prelude::{ListModelExt, SettingsExtManual};
+use gtk::gio::prelude::ListModelExt;
 use gtk::glib::object::CastNone;
 use gtk::prelude::WidgetExt;
 use gtk::subclass::prelude::*;
 use gtk::{ListScrollFlags, SingleSelection, SortListModel, gio, glib};
 
-use crate::config;
 use crate::objects::BssObject;
 use crate::widgets::column_view;
-use columns::*;
+use columns::update_last_seen_label;
 
 mod imp {
     use std::cell::OnceCell;
@@ -184,81 +184,6 @@ impl BssTable {
             .scroll_to(index_of_selected, None, ListScrollFlags::SELECT, None);
     }
 
-    fn setup_columns(&self) {
-        let imp = self.imp();
-
-        // Set up factories for each column
-        imp.color_column.set_factory(Some(&create_color_factory()));
-        imp.ssid_column.set_factory(Some(&create_ssid_factory()));
-        imp.bssid_column.set_factory(Some(&create_bssid_factory()));
-        imp.vendor_column
-            .set_factory(Some(&create_vendor_factory()));
-        imp.signal_column
-            .set_factory(Some(&create_signal_factory()));
-        imp.channel_column
-            .set_factory(Some(&create_channel_factory()));
-        imp.channel_width_column
-            .set_factory(Some(&create_channel_width_factory()));
-        imp.frequency_column
-            .set_factory(Some(&create_frequency_factory()));
-        imp.band_column.set_factory(Some(&create_band_factory()));
-        imp.protocols_column
-            .set_factory(Some(&create_protocols_factory()));
-        imp.amendments_column
-            .set_factory(Some(&create_amendments_factory()));
-        imp.security_column
-            .set_factory(Some(&create_security_factory()));
-        imp.channel_utilization_column
-            .set_factory(Some(&create_channel_utilization_factory()));
-        imp.station_count_column
-            .set_factory(Some(&create_station_count_factory()));
-        imp.max_rate_column
-            .set_factory(Some(&create_max_rate_factory()));
-        imp.uptime_column
-            .set_factory(Some(&create_uptime_factory()));
-
-        // Last Seen column needs the bound_labels collection for refresh tracking
-        imp.last_seen_column
-            .set_factory(Some(&create_last_seen_factory(
-                imp.bound_last_seen_labels.clone(),
-            )));
-    }
-
-    fn setup_column_sorters(&self) {
-        let imp = self.imp();
-
-        imp.ssid_column.set_sorter(Some(&create_ssid_sorter()));
-        imp.bssid_column.set_sorter(Some(&create_bssid_sorter()));
-        imp.vendor_column.set_sorter(Some(&create_vendor_sorter()));
-        imp.signal_column.set_sorter(Some(&create_signal_sorter()));
-        imp.channel_column
-            .set_sorter(Some(&create_channel_sorter()));
-        imp.frequency_column
-            .set_sorter(Some(&create_frequency_sorter()));
-        imp.band_column.set_sorter(Some(&create_band_sorter()));
-        imp.channel_width_column
-            .set_sorter(Some(&create_channel_width_sorter()));
-        imp.protocols_column
-            .set_sorter(Some(&create_protocols_sorter()));
-        imp.amendments_column
-            .set_sorter(Some(&create_amendments_sorter()));
-        imp.security_column
-            .set_sorter(Some(&create_security_sorter()));
-        imp.channel_utilization_column
-            .set_sorter(Some(&create_channel_utilization_sorter()));
-        imp.station_count_column
-            .set_sorter(Some(&create_station_count_sorter()));
-        imp.max_rate_column
-            .set_sorter(Some(&create_max_rate_sorter()));
-        imp.uptime_column.set_sorter(Some(&create_uptime_sorter()));
-        imp.last_seen_column
-            .set_sorter(Some(&create_last_seen_sorter()));
-
-        imp.column_view
-            .sort_by_column(Some(&imp.signal_column), gtk::SortType::Descending);
-    }
-
-    /// Start a timer that periodically refreshes the "Last Seen" labels
     fn start_last_seen_refresh_timer(&self) {
         let bound_labels = self.imp().bound_last_seen_labels.clone();
         let source_id = glib::timeout_add_local(Duration::from_secs(10), move || {
@@ -270,74 +195,10 @@ impl BssTable {
         self.imp().last_seen_timer_id.replace(Some(source_id));
     }
 
-    /// Stop the "Last Seen" refresh timer
     fn stop_last_seen_refresh_timer(&self) {
         if let Some(source_id) = self.imp().last_seen_timer_id.take() {
             source_id.remove();
         }
-    }
-
-    fn setup_column_visibility(&self) {
-        let imp = self.imp();
-        let settings = gio::Settings::new(config::app_id());
-
-        settings
-            .bind("show-bssid-column", &*imp.bssid_column, "visible")
-            .build();
-        settings
-            .bind("show-vendor-column", &*imp.vendor_column, "visible")
-            .build();
-        settings
-            .bind("show-signal-column", &*imp.signal_column, "visible")
-            .build();
-        settings
-            .bind("show-channel-column", &*imp.channel_column, "visible")
-            .build();
-        settings
-            .bind(
-                "show-channel-width-column",
-                &*imp.channel_width_column,
-                "visible",
-            )
-            .build();
-        settings
-            .bind("show-frequency-column", &*imp.frequency_column, "visible")
-            .build();
-        settings
-            .bind("show-band-column", &*imp.band_column, "visible")
-            .build();
-        settings
-            .bind("show-protocols-column", &*imp.protocols_column, "visible")
-            .build();
-        settings
-            .bind("show-amendments-column", &*imp.amendments_column, "visible")
-            .build();
-        settings
-            .bind("show-security-column", &*imp.security_column, "visible")
-            .build();
-        settings
-            .bind("show-max-rate-column", &*imp.max_rate_column, "visible")
-            .build();
-        settings
-            .bind(
-                "show-channel-utilization-column",
-                &*imp.channel_utilization_column,
-                "visible",
-            )
-            .build();
-        settings
-            .bind(
-                "show-stations-column",
-                &*imp.station_count_column,
-                "visible",
-            )
-            .build();
-        settings
-            .bind("show-uptime-column", &*imp.uptime_column, "visible")
-            .build();
-        settings
-            .bind("show-last-seen-column", &*imp.last_seen_column, "visible")
-            .build();
     }
 }
